requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.8.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.36
alembic==1.13.0
psycopg[binary]==3.3.3
greenlet==3.4.0


# Authentication & Security
supabase==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
httpx==0.24.1
aiohttp==3.9.1

# Validation & Data
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code Quality
black==23.12.0
isort==5.13.2
flake8==6.1.0
mypy==1.7.1

# Monitoring
sentry-sdk==1.38.0

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0
